pyproject.toml
[project]
name = "knowledgespaceagent"
version = "0.1.0"
description = "Assistant"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bs4>=0.0.2",
    "docker>=7.1.0",
    "dotenv>=0.9.9",
    "einops>=0.8.1",
    "fastapi>=0.115.12",
    "gcloud>=0.18.3",
    "google-cloud-storage>=3.1.1",
    "ipykernel>=6.29.5",
    "openai>=1.86.0",
    "pydrive>=1.3.1",
    "pydrive2>=1.21.3",
    "requests>=2.32.4",
    "scikit-learn>=1.7.0",
    "sentence-transformers>=5.0.0",
    "tenacity>=9.1.2",
    "torch>=2.7.1",
    "tqdm>=4.67.1",
    "uvicorn>=0.34.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]
